[Refactor] admin valid 추가
